% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass.R
\name{carbon.stock}
\alias{carbon.stock}
\title{Calculating carbon stock in a given area based on tree allometry for rainforest habitat
published in Brown, 1997}
\usage{
carbon.stock(data, plot.size = 100)
}
\arguments{
\item{data}{a data frame or tibble that contains subplot, species, and diamaeter at breast height (DBH, in cm).}

\item{plot.size}{The plot size in m^2. The default value is 100 m^2.}
}
\value{
a tibble containing Subplot name and total carbon content (in ton/ha)
}
\description{
Calculating carbon stock in a given area based on tree allometry for rainforest habitat
published in Brown, 1997
}
\examples{
carbon.stock(Papandayan)
}
