% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity.R
\name{diversity}
\alias{diversity}
\title{Diversity indices}
\usage{
diversity(data)
}
\arguments{
\item{data}{a data frame or tibble that contains subplot, species, and diamaeter at breast height (DBH, in cm).}
}
\value{
a tibble containing subplot and H' value.
}
\description{
Shannon-Wiener diversity index (H'), Simpson's diversity index (D), and Simpson's dominance index measures diversity in a given plot/subplot.
}
\examples{
diversity(Papandayan)
}
