% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic.R
\name{edaphic}
\alias{edaphic}
\title{Calculate soil properties}
\usage{
edaphic(data, d = 7.62, h = 5)
}
\arguments{
\item{data}{a data frame or tibble containing soil measurements with gravimetric methods}

\item{d}{diameter of the core sampler. The default value is 7.62 cm (2 inch).}

\item{h}{height of the core sampler. The default value is 2 cm.}
}
\value{
a tibble of the original input data together with soil bulk density, water content (\%), organic content (\%), and mineral content(\%).
}
\description{
Calculate soil properties
}
\examples{
edaphic(Edaphic_data)
}
