% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic_data.R
\docType{data}
\name{Edaphic_data}
\alias{Edaphic_data}
\title{Measurements of soil physical properties with gravimetric methods}
\format{
\subsection{\code{Edaphic_data}}{

A data frame with 10 rows and 10 columns:
\describe{
\item{Plot}{Plot number}
\item{T}{Soil temperature}
\item{pH}{Soil pH}
\item{Rh}{Soil relative humidity}
\item{Core_Wet}{Soil wet weight for bulk density measurement with core sampler method in gram}
\item{Core_Dry}{Soil dry weight for bulk density measurement with core sampler method in gram}
\item{Cru_Dry}{Soil dry weight from auger bore plus crucible in gram}
\item{Cru_Ash}{Ash weight after furnace from auger bore plus crucible in gram}
\item{Au_Ash}{Ash weight after furnace from auger bore in gram}
\item{Crucible}{Weight of the crucibles in gram}
}
}
}
\source{
Noviana Budianti private repo
}
\usage{
Edaphic_data
}
\description{
A subset of soil physical properties with gravimetric method from Mount Papandayan, West Java, Indonesia.
}
\keyword{datasets}
