% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVI.R
\name{IVI}
\alias{IVI}
\title{Analyze IVI from DBH data}
\usage{
IVI(data, plot.size = 100, n.subplot = 5)
}
\arguments{
\item{data}{a data frame or tibble that contains subplot, species, and diamaeter at breast height (DBH, in cm)}

\item{plot.size}{The plot size in m^2. The default value is 100 m^2.}

\item{n.subplot}{The number of subplot. The default value is 5.}
}
\value{
a tibble containing species name, relative frequency, relative density, relative coverage, and importance value index.
}
\description{
Analyze IVI from DBH data
}
