% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVI_plot.R
\name{IVI_plot}
\alias{IVI_plot}
\title{Plot of importance value index}
\usage{
IVI_plot(data)
}
\arguments{
\item{data}{a tibble containing importance value index generated by function IVI()}
}
\value{
A plot (ggplot) of importance value index and its components, ordered descending
}
\description{
Plot of importance value index
}
