% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic.R
\name{mineral}
\alias{mineral}
\title{Soil mineral content}
\usage{
mineral(dry, ash, crucible, data = NULL)
}
\arguments{
\item{dry}{soil dry weight in gram.}

\item{ash}{ash weight in gram.}

\item{crucible}{crucible weight in gram.}

\item{data}{data frame containing edaphic data, measured by gravimetric methods.}
}
\value{
mineral content in soil (in percent).
}
\description{
Soil mineral content
}
\examples{
mineral(x, y, z, data = df)
}
