% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Papandayan.R
\docType{data}
\name{Papandayan}
\alias{Papandayan}
\title{Analysis of tree vegetation in Papandayan}
\format{
\subsection{\code{Papandayan}}{

A data frame with 77 rows and 3 columns:
\describe{
\item{Subplot}{Subplot number}
\item{Species}{Speceis binomial nomenclature name}
\item{DBH}{Diameter at breast height data in cm}
}
}
}
\source{
Noviana Budianti private repo
}
\usage{
Papandayan
}
\description{
A subset of diameter at breast height data from Mount Papandayan, West Java, Indonesia.
}
\keyword{datasets}
