% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rho.R
\docType{data}
\name{Rho}
\alias{Rho}
\title{Database of wood density}
\format{
\subsection{\code{Rho}}{

A data frame with 33 rows and 4 columns:
\describe{
\item{Family}{Family name according to binomial nomenclature system}
\item{Local_Name}{Local name of the tree species}
\item{Species}{Speceis binomial nomenclature name}
\item{Rho}{Wood density in g.cm-3}
}
}
}
\source{
Noviana Budianti private repo
}
\usage{
Rho
}
\description{
A database of wood density (g.cm-3) of various tree species in Mount Papandayan,
West Java, Indonesia.
}
\keyword{datasets}
