% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass.R
\name{biomass}
\alias{biomass}
\title{Biomass calculation according to tree allometry for different reinforest habitat.}
\usage{
biomass(data, method = c("dry", "moist", "wet"))
}
\arguments{
\item{data}{a data frame or tibble that contains subplot, species, and diamaeter at breast height (DBH, in cm).}

\item{method}{a climate parameter that detemines the allometry used, \emph{"dry"}, \emph{"moist"}, and \emph{"wet".}}
}
\value{
a tibble containing species name, total diamater at breast height (in cm), and total biomass (in kg).
}
\description{
Biomass calculation according to tree allometry for different reinforest habitat.
}
\examples{
biomass(Papandayan, "moist")
}
