% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass.R
\name{biomass}
\alias{biomass}
\title{Biomass calculation according to tree allometry for moist habitat in Brown, 1997}
\usage{
biomass(data)
}
\arguments{
\item{data}{a data frame or tibble that contains subplot, species, and diamaeter at breast height (DBH, in cm).}
}
\value{
a tibble containing species name, total diamater at breast height (in cm), and total biomass (in kg).
}
\description{
Biomass calculation according to tree allometry for moist habitat in Brown, 1997
}
\examples{
biomass(Papandayan)
}
