% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic.R
\name{bulk}
\alias{bulk}
\title{Soil bulk density}
\usage{
bulk(dry, data = NULL, d = 7.62, h = 5)
}
\arguments{
\item{dry}{soil dry weight in gram.}

\item{data}{data frame containing edaphic data, measured by gravimetric methods.}

\item{d}{diameter of the core sampler in cm. The default value is 7.62 cm (3").}

\item{h}{height of the core sampler in cm. The default value is 5 cm.}
}
\value{
Soil bulk density in gram.cm^-3
}
\description{
Soil bulk density
}
\examples{
bulk(i, data = df)
}
