% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass.R
\name{carbon}
\alias{carbon}
\title{Calculating carbon content in tree based on biomass calculated with tree allometry for moist habitat in Brown, 1997}
\usage{
carbon(data, C = 0.5)
}
\arguments{
\item{data}{a data frame or tibble that contains subplot, species, and diamaeter at breast height (DBH, in cm).}

\item{C}{Coefficient for carbon content in a tree.}
}
\value{
a tibble containing species name, total diamater at breast height (in cm), total biomass (in kg), and total carbon content (in kg)
}
\description{
Calculating carbon content in tree based on biomass calculated with tree allometry for moist habitat in Brown, 1997
}
\examples{
carbon(Papandayan)
}
