% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass.R
\name{est.height}
\alias{est.height}
\title{Estimating height from diameter at breast height measurement}
\usage{
est.height(data)
}
\arguments{
\item{DBH}{Measurement of diameter at breast height (cm)}
}
\value{
tree estimated height (in m)
}
\description{
Estimating height from diameter at breast height measurement
}
\examples{
est.height(5)
}
