% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic.R
\name{organic}
\alias{organic}
\title{Soil organic content}
\usage{
organic(dry, ash, crucible, data = NULL)
}
\arguments{
\item{dry}{soil dry weight in gram.}

\item{ash}{ash weight in gram.}

\item{crucible}{crucible weight in gram.}

\item{data}{data frame containing edaphic data, measured by gravimetric methods.}
}
\value{
organic content in soil (in percent).
}
\description{
Soil organic content
}
\examples{
organic(x, y, z, data = df)
}
