% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic.R
\name{pore}
\alias{pore}
\title{Soil porosity}
\usage{
pore(
  bulk,
  type = c("low organic", "medium organic", "high organic"),
  data = NULL
)
}
\arguments{
\item{bulk}{bulk density in gram.cm^-3.}

\item{type}{type of soil. Determining the average porosity in the given type, expressed in gram.cm^-3. "low organic" 2,7 gram.cm^-3, "medium organic" 2,65 gram.cm^-3, "high organic" 2,6 gram.cm^-3.}

\item{data}{data frame containing edaphic data, measured by gravimetric methods.}
}
\value{
Soil porosityin gram.cm^-3.
}
\description{
Soil porosity measures the density of pore space in a soil matrix, i.e. the opposite of soil bulk density.
}
\examples{
pore(j, "high organic", data = df)
}
