% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edaphic.R
\name{water}
\alias{water}
\title{Soil water content}
\usage{
water(wet, dry, container, data = NULL)
}
\arguments{
\item{wet}{soil wet weight in gram.}

\item{dry}{soil dry weight in gram.}

\item{container}{soil container weight in gram.}

\item{data}{data frame containing edaphic data, measured by gravimetric methods.}
}
\value{
water content in soil in percent.
}
\description{
Soil water content
}
\examples{
water(x, y, data = df)
}
